reset peer instead of close
When you call close it will send a close request to the peer,
if the peer does not respond then it stays in the closing state
and you are unable to open a new connection or create a new
peer. By calling reset peer we still send a close request but we also
clear the peer.